Support using the Storefront private access token
This allows for using [the Storefront private access token][1] with the
library's Storefront GraphQL client. Since this client is probably going
to be running on a server where the token can be secured, using the
private access token should be preferred. This commit doesn't make using
the private token the default to avoid an immediate breaking change. We
want to give clients time to see the deprecation and update their
scripts.

Closes #1293

[1]: https://shopify.dev/docs/api/usage/authentication#getting-started-with-private-access