Extend webhook registration to support filters
With this change, we can register webhooks specifying the filter
parameter that allows to define some conditions on the topic fields
to reduce the number of webhooks received.

Other than adding this feature for already possible registrations, this
extends the number of webhooks that can be registered with this library.

In fact, there are some webhook registrations that require the filter
parameter to be passed to allow the registrationi (this is partly an
assumption because I couldn't find proper documentation around this
topic). An example is METAOBJECTS_CREATE, which, without the filter
param set,

```
mutation webhookSubscriptionCreate($topic: WebhookSubscriptionTopic!, $webhookSubscription: WebhookSubscriptionInput!) {
  webhookSubscriptionCreate(topic: $topic, webhookSubscription: $webhookSubscription) {
    webhookSubscription {
    	topic
      filter
    }
    userErrors {
      field
      message
    }
  }
}

{
  "topic": "METAOBJECTS_CREATE",
  "webhookSubscription": {
    "callbackUrl": "https://test5.com"
  }
}

```

gives the following error:

```
{
  "data": {
    "webhookSubscriptionCreate": {
      "webhookSubscription": null,
      "userErrors": [
        {
          "field": [
            "webhookSubscription"
          ],
          "message": "The specified filter is invalid, please ensure you specify the field(s) you wish to filter on."
        }
      ]
    }
  }
}
```

While with the filter on, works as expected.

---

This change has been made available to 2024-07 and 2024-10, which are
the versions with the filter parameter available, according to the
GraphQL documentation.

---

Refs:
- https://shopify.dev/docs/apps/build/webhooks/customize/filters
- https://shopify.dev/docs/api/admin-graphql/2024-07/input-objects/WebhookSubscriptionInput
- https://shopify.dev/docs/api/admin-graphql/2024-10/input-objects/WebhookSubscriptionInput